packages/harpguru-core/src/components/callback-on-sourceglobalprops/utils/reduce-harpstrata-to-fulldegreematrix/reduce-harpstrata-to-fulldegreematrix.ts
import type { HarpStrata } from 'harpstrata'
import type { HarpFaceMatrix, Degree } from 'harpparts'

import { isMatchHighOrderTuples } from '../../../../utils'
import { doSparceIdedObjectMatricesMatch } from '../../../../packages/do-sparce-ided-object-matrices-match'

export const reduceHarpStrataToFullDegreeMatrix = (
  prevDegreeMatrix: readonly [HarpFaceMatrix<Degree>, HarpFaceMatrix<Degree>],
  harpStrata: HarpStrata
): readonly [HarpFaceMatrix<Degree>, HarpFaceMatrix<Degree>] => {
  const { degreeMatrix: nextDegreeMatrix } = harpStrata
  if (
    isMatchHighOrderTuples(
      doSparceIdedObjectMatricesMatch,
      prevDegreeMatrix,
      nextDegreeMatrix
    )
  )
    return prevDegreeMatrix
  return nextDegreeMatrix
}
